[emscripten] Build a packageable directory for emdawnwebgpu
This creates a new output directory `emdawnwebgpu_pkg` in the build
root. This contains exactly the files needed to link emdawnwebgpu, and a
"port" file that can be used via `emcc --use-port`.

This is a non-standard use of port files; usually they're stored in the
Emscripten source tree, and download and compile a source dependency.
But it so happens they also provide all the hooks needed for us to
insert include paths and link flags, without having to integrate into
the user's build/meta-build system.

The structure of the package, at out/*/emdawnwebgpu_pkg, is:

emdawnwebgpu_pkg
├── emdawnwebgpu.port.py
├── README.md
├── webgpu
│   ├── include
│   │   └── webgpu
│   │       └── webgpu.h
│   └── src
│       ├── library_webgpu_enum_tables.js
│       ├── library_webgpu_generated_sig_info.js
│       ├── library_webgpu_generated_struct_info.js
│       ├── library_webgpu.js
│       ├── LICENSE
│       ├── webgpu-externs.js
│       └── webgpu.cpp
└── webgpu_cpp
    ├── include
    │   ├── dawn
    │   │   └── webgpu_cpp_print.h
    │   └── webgpu
    │       ├── webgpu_cpp_chained_struct.h
    │       ├── webgpu_cpp.h
    │       └── webgpu_enum_class_bitmasks.h
    └── LICENSE

Bug: 371024051
Change-Id: I60e0883ce0e44be9413dca1439cdb30382571cdf