Publish emdawnwebgpu in a GitHub release
This runs a github action which:
- Builds emdawnwebgpu_pkg
- Saves the version info into emdawnwebgpu_pkg/VERSION.txt
- Zips emdawnwebgpu_pkg
- Creates a git tag VERSION
- Creates a release for that tag with emdawnwebgpu_pkg-VERSION.zip

If someday we publish additional build artifacts, they can be published
as additional files in the same release, alongside this one (e.g. native
binaries, or CMake configs that automatically select between Dawn and
emdawnwebgpu based on build target).

Note the Git tags will be created via the GitHub API as they are
required in order to publish releases on GitHub. These probably won't
get synced into googlesource.com, but  hopefully the sync won't try to
delete the tags. I think GitHub will prevent that, but that could cause
the sync to stop working.

Bug: 371024051
Change-Id: I59683a8ec40e0db940e35c9e1ea8c7e678686977